fix: cron middleware

+++ app/memento/cron/cleanup/route.ts
@@ -55,12 +55,12 @@ async function deleteOneWeekOldMemento() {
 }
 
 export async function GET(request: Request) {
-    // const authHeader = request.headers.get('authorization');
-    // if (authHeader !== `Bearer ${process.env.CRON_SECRET}`) {
-    //     return new Response('Unauthorized', {
-    //         status: 401,
-    //     });
-    // }
+    const authHeader = request.headers.get('authorization');
+    if (authHeader !== `Bearer ${process.env.CRON_SECRET}`) {
+        return new Response('Unauthorized', {
+            status: 401,
+        });
+    }
 
     try {
         const data = await deleteOneWeekOldMemento();
